+++ src/modules/batch/index.ts
@@ -15,6 +15,8 @@ class Batch implements IBatch {
   claimHostUrl: string
   campaignData: TCampaignItem
   signerKey: string
+  apiKey: string
+  chainId: number
 
   constructor (
     batchId: string,
@@ -25,7 +27,9 @@ class Batch implements IBatch {
     campaignData: TCampaignItem,
     signerKey: string,
     campaignSig: string,
-    apiHost: string
+    apiHost: string,
+    apiKey: string,
+    chainId: number
   ) {
     this.batchId = batchId
     this.data = data
@@ -36,6 +40,8 @@ class Batch implements IBatch {
     this.claimHostUrl = claimHostUrl
     this.campaignData = campaignData
     this.signerKey = signerKey
+    this.apiKey = apiKey
+    this.chainId = chainId
   }
 
   addLinks: TAddLinks = async (
@@ -66,6 +72,7 @@ class Batch implements IBatch {
     if (!transformedAssets) { return alert('Error with assets') }
     return await batchesApi.addLinks(
       this.apiHost,
+      this.apiKey,
       this.campaignSig,
       this.campaignData.campaign_id,
       this.batchId,
@@ -73,17 +80,26 @@ class Batch implements IBatch {
     )
   }
 
-  getLinks: TGetLinks = () => {
+  getLinks: TGetLinks = (
+    linkPattern
+  ) => {
     if (!this.claimLinks) {
       return []
     }
     return this.claimLinks.map(link => {
       const encryptedClaimCode = link.encrypted_claim_code
       const claimCode = crypto.decrypt(encryptedClaimCode, this.encryptionKey)
+
+      let finalLink = `${this.claimHostUrl}/#/redeem/${claimCode}`
+      if (linkPattern) {
+        finalLink = linkPattern.replace('<CODE>', claimCode)
+          .replace('<CHAIN_ID>', String(this.chainId))
+      }
+      const sourceParam = finalLink.includes('?') ? `&src=d` : '?src=d'
       return {
         linkId: link.link_id,
         claimCode,
-        claimLink: `${this.claimHostUrl}/#/redeem/${claimCode}?src=d`
+        claimLink: `${finalLink}${sourceParam}`
       }
     })
   }