+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Linkdrop SDK
 
+
+## 2.1.3
+- updates for src=d param (getLinks method of batch)
+
+## 2.1.2
+- updates for claimHostUrl param (getLinks method of batch)
+
+## 2.1.1
+- added testnets apiUrl
+
+## 2.1.0
+- added apiKey as param for LinkdropSDK 
+
 ## 2.0.9
 - added src param to link (src=d)